LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A5_T2-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Label 'IN_DO_FUNC' is not defined
S12.7_A5_T2

PASS Expected parsing failure

TEST COMPLETE
